requirements.txt
# Ultimate Math Agent - Dependencies
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0

# LLM Provider SDKs
openai>=1.50.0
anthropic>=0.35.0
google-generativeai>=0.8.0
httpx>=0.27.0

# Math & Symbolic Computation
sympy>=1.13.0
numpy>=1.26.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0

# Web UI
gradio>=4.44.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
